+++ Products/PluggableAuthService/plugins/ZODBGroupManager.py
@@ -356,7 +356,7 @@ class ZODBGroupManager(BasePlugin):
     def manage_removeGroups(self, group_ids, RESPONSE=None, REQUEST=None):
         """ Remove one or more groups via the ZMI.
         """
-        group_ids = filter(None, group_ids)
+        group_ids = [_f for _f in group_ids if _f]
 
         if not group_ids:
             message = 'no+groups+selected'